--- src/main/java/com/BrowserStack/LocalException.java
@@ -1,7 +1,0 @@
-package com.BrowserStack;
-
-class LocalException extends Exception {
-	LocalException(String message){
-		super(message);
-	}
-}
