+++ src/main/java/com/browserstack/local/LocalException.java
@@ -1,0 +1,8 @@
+package com.browserstack.local;
+
+class LocalException extends Exception {
+
+    LocalException(String message) {
+        super(message);
+    }
+}
