--- src/test/java/com/BrowserStack/BrowserStackLocalTest.java
@@ -1,126 +1,0 @@
-package com.BrowserStack;
-
-import java.util.HashMap;
-import org.junit.*;
-import static org.junit.Assert.*;
-
-public class BrowserStackLocalTest {
-  Local l; 
-  HashMap<String, String> options;
-
-  @Before
-  public void setUp() throws Exception {
-    l = new Local();
-    options = new HashMap<String, String>();
-    options.put("key", System.getenv("BROWSERSTACK_ACCESS_KEY"));
-  }
-
-  @Test
-  public void testIsRunning() throws Exception {
-    assertFalse(l.isRunning());
-    l.start(options);
-    assertTrue(l.isRunning());
-  }
-
-  @Test
-  public void testMultipleBinary() throws Exception {
-    l.start(options);
-    assertTrue(l.isRunning());
-    Local l2 = new Local();
-    try {
-      l2.start(options);
-    }
-    catch(LocalException e){
-      assertFalse(l2.isRunning());
-    }
-  }
-
-  @Test
-  public void testEnableVerbose() throws Exception {
-    options.put("v", "");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-vvv"));
-  }
-
-  @Test
-  public void testSetFolder() throws Exception {
-    options.put("f", "/var/html");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-f"));
-    assertTrue(l.command.contains("/var/html"));
-  }
-
-  @Test
-  public void testEnableForce() throws Exception {
-    options.put("force", "");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-force"));
-  }
-
-  @Test
-  public void testEnableOnly() throws Exception {
-    options.put("only", "");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-only"));
-  }
-
-  @Test
-  public void testEnableOnlyAutomate() throws Exception {
-    options.put("onlyAutomate", "");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-onlyAutomate"));
-  }
-
-  @Test
-  public void testEnableForceLocal() throws Exception {
-    options.put("forcelocal", "");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-forcelocal"));
-  }
-
-  @Test
-  public void testSetLocalIdentifier() throws Exception {
-    options.put("localIdentifier", "abcdef");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-localIdentifier"));
-    assertTrue(l.command.contains("abcdef"));
-  }
-
-  @Test
-  public void testSetProxy() throws Exception {
-    options.put("proxyHost", "localhost");
-    options.put("proxyPort", "8080");
-    options.put("proxyUser", "user");
-    options.put("proxyPass", "pass");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("-proxyHost"));
-    assertTrue(l.command.contains("localhost"));
-    assertTrue(l.command.contains("-proxyPort"));
-    assertTrue(l.command.contains("8080"));
-    assertTrue(l.command.contains("-proxyUser"));
-    assertTrue(l.command.contains("user"));
-    assertTrue(l.command.contains("-proxyPass"));
-    assertTrue(l.command.contains("pass"));
-  }
-
-  @Test
-  public void testSetHosts() throws Exception {
-    options.put("hosts", "localhost,8000,0");
-    options.put("onlyCommand", "true");
-    l.start(options);
-    assertTrue(l.command.contains("localhost,8000,0"));
-  }
-
-  @After
-  public void tearDown() throws Exception {  
-    l.stop();
-  }  
-}
